backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-mail==0.9.1
python-dotenv==1.0.0
gunicorn==21.2.0
google-generativeai==0.3.2
